Use script in the mypy repository to build wheels
This removes inline configuration and moves it to a script.
This makes it easier to have more complex logic during the build,
and also we can more easily test the build locally.

Also there's less need to understand the details of how GitHub actions
are configured.

Related to python/mypy#10074.